+++ frontend/package.json
@@ -11,7 +11,7 @@
     "@testing-library/react": "^16.2.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.8.4",
-    "bootstrap": "^5.3.4",
+    "bootstrap": "^5.3.5",
     "qrcode.react": "^4.2.0",
     "react": "^19.1.0",
     "react-bootstrap": "^2.10.9",
