+++ frontend/src/pages/Promotions/assets/search.png
