+++ frontend/src/pages/Promotions/assets/supermarket.avif
