+++ frontend/src/pages/Promotions/components/PopUp.css
