+++ frontend/src/pages/Promotions/components/PopUp.jsx
