+++ frontend/src/pages/Promotions/components/Promotions.css
