+++ frontend/src/pages/Promotions/index.js
@@ -1,0 +1,8 @@
+import ManagerPromotions from "./components/Manager/ManagerPromotions";
+import RegularPromotions from "./components/Regular/RegularPromotions";
+
+const Promotions = () => {
+  return <RegularPromotions />;
+}
+
+export default Promotions;
